Final Project Code
This is a push containing the final code from me and partners finished project.

get_input.ino - Arduino code for MOC. This is responsible for getting input from the mobile application. Information is sent to the device from the application and received via bluetooth. From here, that signal is sent to the motor_controller code via a serial connection to control the DC motors.

 motor_control.ino - Arduino code for ACU. This is responsible for receiving signals from get_input and executing their commands. If the user selected Autonomous mode in the mobile app, the MOC subsystem lets the ACU know via serial connection, and the motor_control.ino handles obstacle avoidance. In contrast, if Manual mode has been selected, the motor_control.ino will execute signals received from the other subsystem in order to turn let/right and move foward/stop.